Finished Delete Modal

+++ lib/deleteFile.ts
@@ -1,9 +1,11 @@
 import axios from "axios";
 
-export const deleteValue = async (id: string) => {
-  await axios.delete("/api/uploadthing", {
-    data: {
-        url: id,
-    },
-  });
+export const deleteValue = async (id?: string) => {
+  if (id) {
+    await axios.delete("/api/uploadthing", {
+      data: {
+          url: id,
+      },
+    });
+  }
 };
